Static resources
Summary:
Provides a means for Clams to serve static resources. Builds on top of
Compojure, just like the rest of the routing. Production apps will want
to serve static resources directly from a server like nginx (rather than
from the Java container/http-kit). This diff doesn't address that use case
but rather just the development case, so that remains a todo for the future.

Test Plan:
Inspection. Unfortunately this kind of interfacing with Compojure isn't very
testable and I haven't built in an integration test system yet. That's also
a todo.

Reviewers: #standard_treasury, david

Reviewed By: #standard_treasury, david

Subscribers: david

Differential Revision: http://phabricator.corp.standardtreasury.com/D3812