Config modes
Summary:
Introduces the ability to set a config mode. If the app is passed a value
in the environment variable `CLAMS_ENV`, then the corresponding config file
will be loaded and merged. The config file must be located at
`/resources/conf/$(CLAMS_ENV).edn`. It has the highest merge priority with
the exception of other environment variables. The intent is to use this feature
to easily switch between dev and prod (etc.) configs at run-time.

Test Plan:
  lein test

Reviewers: #standard_treasury, ctdean

Reviewed By: #standard_treasury, ctdean

Differential Revision: http://phabricator.corp.standardtreasury.com/D3818