Allow CLAMS_ENV to be set using a Java system property
Summary:
In general I dislike Java system properties. They seem to be a gratuitous take
on environment variables. And they're mutable--so what's the point?
Unfortunately Leiningen allows them to be set per-profile, while the real env
vars get no love.

In order for apps to support a Clams "testing" config, without having to resort
to other libs or hacks, we need a way to set `CLAMS_ENV=test` from the
Leiningen test runner. This diff extends conf to look for a property named
`clams.env` that can be used in lieu of the equivalent env var.

Test Plan:
  lein test

Reviewers: #standard_treasury, ctdean

Reviewed By: #standard_treasury, ctdean

Subscribers: ctdean

Differential Revision: http://phabricator.corp.standardtreasury.com/D3841