User-defined parameter binding
Summary:
A much-requested feature in Clams is the ability to access arbitrary parts of
the request map that exist outside of the `:params` key and bind them to
controller params (e.g. see D3895). For various reasons (mainly because the
syntax is very complex and unwieldy), Clams has elected to avoid the typical
Clojure binding/destructuring approach and implement its own system that's
softer and cuddlier (i.e. see `clams.params`).

This diff is a proposal for how to implement that feature. It adds the ability
to pass an arbitrary function where you would otherwise pass one of the schema
specifiers from `clams.params`. When the request parameters are validated,
those names with have functions as their schema are skipped, and instead are
simply mapped onto the result of applying their associated function on the full
request map.

In this way, Clams apps are free to define arbitrary methods for extracting
non-standard request params, in particular those that aren't stored under the
`:params` key.

See also D3907 for an example of this in action.

Test Plan:
  lein test

Reviewers: #standard_treasury, ctdean

Reviewed By: #standard_treasury, ctdean

Subscribers: ctdean

Differential Revision: http://phabricator.corp.standardtreasury.com/D3906