[middleware] Reordered middleware to prevent unnesting of JSON
Summary:
ring-json has some interaction problems with wrap-nested-params; this
re-orders our middleware to avoid that particular use case and adds a
test for it.

In particular, if `wrap-nested-params` is called after the json middleware,
it will flatten any nested objects contained within vectors by grabbing the
first item in the vector and replacing the vector with that. This basically
makes json arrays unusable and is obviously an undesirable bug.

This revision also ended up removing `wrap-json-body` since having both
`wrap-json-body` and `wrap-json-params` was deemed both un-necessary
and problematic.

To be specific about the above: if `wrap-json-body` is evaluated first, it
replaces the `:body` key of the request map with a map instead of an InputStream
instance. This then causes the next piece of middleware that tries to slurp
from the InputStream to throw an exception. For this reason, `wrap-json-body`
cannot happen before `wrap-json-params` (and I would imagine some list of other
middleware).

If `wrap-json-params` is evaluated before `wrap-json-body`, the json content is
correctly parsed and assoc'd into the request map's `:json-params` key, and merged
into the request map's `:params` key. However, for reasons I have not been able to
ascertain, it ends up leaving the `:body` key as nil.

So you're at least partially damned, no matter what you do. Given that, and given our
preference for json content to sit in either `:json-params` or `:params` for other reasons,
I've also removed `wrap-json-body` from our list of middleware.

Test Plan: CircleCI

Reviewers: ctdean, keith, mike, jim

Reviewed By: jim

Subscribers: jim

Differential Revision: http://phabricator.corp.standardtreasury.com/D4045